feat(zones): support for editing zones
